Move _TrackNVDAInitialization to NVDAState
