+++ source/NVDAState.py
@@ -44,3 +44,27 @@ def _getExitCode() -> int:
 
 def _setExitCode(exitCode: int) -> None:
 	globalVars.exitCode = exitCode
+
+
+class _TrackNVDAInitialization:
+	"""
+	During NVDA initialization,
+	core._initializeObjectCaches needs to cache the desktop object,
+	regardless of lock state.
+	Security checks may cause the desktop object to not be set if NVDA starts on the lock screen.
+	As such, during initialization, NVDA should behave as if Windows is unlocked,
+	i.e. winAPI.sessionTracking._isLockScreenModeActive should return False.
+	"""
+
+	_isNVDAInitialized = False
+	"""When False, _isLockScreenModeActive is forced to return False.
+	"""
+
+	@staticmethod
+	def markInitializationComplete():
+		assert not _TrackNVDAInitialization._isNVDAInitialized
+		_TrackNVDAInitialization._isNVDAInitialized = True
+
+	@staticmethod
+	def isInitializationComplete() -> bool:
+		return _TrackNVDAInitialization._isNVDAInitialized
